gets correct primary title for object #106
