+++ lib/transforms/get-values.js
@@ -1,4 +1,5 @@
 const getNestedProperty = require('../nested-property');
+const getPrimaryValue = require('../get-primary-value');
 
 module.exports = {
   getMadeDate (data) {
@@ -27,5 +28,9 @@ module.exports = {
     if (item.type === 'documents') {
       return getNestedProperty(item, 'attributes.level.value') === 'fonds' ? 'archive' : 'documents';
     }
+  },
+
+  getTitle (item) {
+    return getPrimaryValue(item.attributes.title) || item.attributes.summary_title;
   }
 };
