+++ lib/transforms/json-to-html-data.js
@@ -2,7 +2,7 @@ const getNestedProperty = require('../nested-property');
 const getValues = require('./get-values');
 
 module.exports = (resource) => {
-  const title = getTitle(resource.data);
+  const title = getValues.getTitle(resource.data);
   const type = getType(resource.data);
   const fact = getFacts(resource);
   const related = getRelated(resource);
@@ -14,10 +14,6 @@ module.exports = (resource) => {
 
   return {title, type, fact, related, description, details, displayLocation, system, level};
 };
-
-function getTitle (data) {
-  return data.attributes.summary_title;
-}
 
 function getType (data) {
   return data.type;
