+++ lib/transforms/search-results-to-template-data.js
@@ -3,6 +3,7 @@ const formatNumber = require('format-number')();
 const Pagination = require('../pagination');
 const getValues = require('./get-values');
 const getNestedProperty = require('../nested-property');
+const getPrimaryValue = require('../get-primary-value');
 
 module.exports = (params, results, config) => {
   config = config || {};
@@ -45,7 +46,7 @@ module.exports = (params, results, config) => {
 function createResult (item) {
   return {
     type: item.type,
-    title: getTitle(item),
+    title: getValues.getTitle(item),
     category: getCategory(item),
     date: getDate(item),
     figure: getFigure(item),
@@ -55,10 +56,6 @@ function createResult (item) {
     isOnDisplay: getValues.getDisplayLocation(item),
     level: getValues.getDocumentLevel(item)
   };
-}
-
-function getTitle (item) {
-  return getPrimaryValue(item.attributes.title) || item.attributes.summary_title;
 }
 
 function getCategory (item) {
@@ -97,19 +94,6 @@ function getFigCaption (item) {
   } else if (item.type === 'people') {
     return getBriefBiography(item);
   }
-}
-
-function getPrimaryValue (array) {
-  if (Object.prototype.toString.call(array) === '[object String]') {
-    return array;
-  }
-
-  if (!Array.isArray(array) || !array.length) {
-    return null;
-  }
-
-  const primary = array.find((d) => d.primary);
-  return primary ? primary.value : array[0].value;
 }
 
 function getBriefBiography (item) {
